Fix clear value array in textoverlay
According to section 7.4. Render Pass Commands of the Vulkan spec,
VkRenderPassBeginInfo::pClearValues is indexed by attachment number.
Since the second attachment in the renderpass will be cleared, a clear
value array of length two is required, and the second element of that
array must be assigned the desired clear value.