Update WM_CLASS property on Linux using xcb
This commit updates the WM_CLASS property to display a better
title than "Unknown" for each example in the dash and application menu.

Update WM_CLASS property on Linux using xcb

This commit updates the WM_CLASS property to display a better
title than "Unknown" for each example in the dash and application menu.