Update computeshader.cpp
There is a logical error that cause the first frame render error.  Since the graphic rendering need to sample the compute shader 's output image, so the compute command need to run  before the graphic rendering. If not, the graphic rendering will sample the gabage data (computer's output image) for the first frame and present on the display screen.This issue can be easy to be reproduced when the GPU device has a lower performance.