Revert recent change to assets install for cross-compile
Since existing textures and models are still installed at the root, it
didn't have any noticeable effect other than to create an empty folder.
The change was made because it looked like it should be updated like
the shaders folder, but it wasn't solving any problem, so switching
it back seems the best.